Resolve click not working
Tried to run this game after two years and found out the mouse does not work. I fixed it and put the game back to playable state. But looking through the code I would rewrite 90% of the game, mainly the animations and the way entities are handled.
Maybe one day remastered version as lot of developers seem to like the idea nowdays? XD